+++ mapper/src/main/java/org/grnet/cat/mappers/registry/MetricDefinitionMapper.java
@@ -1,8 +1,10 @@
 package org.grnet.cat.mappers.registry;
 
 import org.apache.commons.lang3.StringUtils;
+import org.grnet.cat.dtos.registry.MetricDefinitionExtendedResponse;
 import org.grnet.cat.dtos.registry.MetricDefinitionResponseDto;
 import org.grnet.cat.entities.registry.MetricDefinitionJunction;
+import org.grnet.cat.mappers.registry.metric.MetricMapper;
 import org.mapstruct.IterableMapping;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -11,7 +13,7 @@ import org.mapstruct.factory.Mappers;
 
 import java.util.List;
 
-@Mapper(imports = {StringUtils.class, java.sql.Timestamp.class, java.time.Instant.class})
+@Mapper(imports = {StringUtils.class, java.sql.Timestamp.class, java.time.Instant.class}, uses = {MetricMapper.class, TypeBenchmarkMapper.class})
 public interface MetricDefinitionMapper {
 
     MetricDefinitionMapper INSTANCE = Mappers.getMapper(MetricDefinitionMapper.class);
@@ -24,5 +26,25 @@ public interface MetricDefinitionMapper {
     @Mapping(target = "typeBenchmarkId",  expression = "java(metricDefinitionJunction.getTypeBenchmark().getId())")
     @Mapping(target = "motivationId", expression = "java(metricDefinitionJunction.getMotivation().getId())")
     MetricDefinitionResponseDto metricDefinitionToResponseDto(MetricDefinitionJunction metricDefinitionJunction);
+
+    @IterableMapping(qualifiedByName = "mapToExtendedResponse")
+    List<MetricDefinitionExtendedResponse> metricDefinitionToExtendedResponses(List<MetricDefinitionJunction> metricDefinitionJunctions);
+
+    @Named("mapToExtendedResponse")
+    @Mapping(target = "metricId", expression = "java(metricDefinitionJunction.getMetric().getId())")
+    @Mapping(target = "metricLabel", expression = "java(metricDefinitionJunction.getMetric().getLabelMetric())")
+    @Mapping(target = "metricDescription", expression = "java(metricDefinitionJunction.getMetric().getDescrMetric())")
+    @Mapping(target = "typeAlgorithmId", expression = "java(metricDefinitionJunction.getMetric().getTypeAlgorithm().getId())")
+    @Mapping(target = "typeAlgorithmLabel", expression = "java(metricDefinitionJunction.getMetric().getTypeAlgorithm().getLabelAlgorithmType())")
+    @Mapping(target = "typeMetricId", expression = "java(metricDefinitionJunction.getMetric().getTypeMetric().getId())")
+    @Mapping(target = "typeMetricLabel", expression = "java(metricDefinitionJunction.getMetric().getTypeMetric().getLabelTypeMetric())")
+    @Mapping(target = "typeBenchmarkId", expression = "java(metricDefinitionJunction.getTypeBenchmark().getId())")
+    @Mapping(target = "typeBenchmarkLabel", expression = "java(metricDefinitionJunction.getTypeBenchmark().getLabelBenchmarkType())")
+    @Mapping(target = "typeBenchmarkDescription", expression = "java(metricDefinitionJunction.getTypeBenchmark().getDescBenchmarkType())")
+    @Mapping(target = "typeBenchmarkPatter", expression = "java(metricDefinitionJunction.getTypeBenchmark().getPattern())")
+    @Mapping(target = "motivationId", expression = "java(metricDefinitionJunction.getMotivation().getId())")
+    MetricDefinitionExtendedResponse metricDefinitionToExtendedResponse(MetricDefinitionJunction metricDefinitionJunction);
+
+
 
 }